requirements.txt
# Legal Document Management AI Agent - Requirements
# Python 3.10+ required

# Microsoft Agent Framework (PREVIEW - requires --pre flag)
agent-framework-azure-ai>=0.1.0

# Azure SDK dependencies
azure-identity>=1.16.0
azure-search-documents>=11.6.0
azure-ai-documentintelligence>=1.0.0
azure-ai-formrecognizer>=3.3.3
azure-ai-inference>=1.0.0b9
azure-core>=1.30.0
openai>=1.12.0

# Configuration management
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.5
asyncio>=3.4.3

# Logging and monitoring
azure-monitor-opentelemetry>=1.5.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0

# Web framework (for API deployment)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# MCP (Model Context Protocol) support
mcp>=1.0.0

# Data handling
pandas>=2.2.0
numpy>=1.26.0

# Security
cryptography>=42.0.0

# Testing (optional)
pytest>=8.1.0
pytest-asyncio>=0.23.0
